Update README.md
Adding the first section - Documentation. We can't tell computers (or people) what to do if we can't first write it down clearly. We expect our agents to perform in compliance with the objectives of the organization. These objectives require clear communication in the form of written word. Thus spoke the machine.